Youth Quality of Life Instrument - Deafness and Hard-of-Hearing Module (YQOL-DHH)
The Research Version (YQOL-DHH) consists of 32 perceptual (known only to the individual) items. The initial validation of the instrument was done with adolescents ages 11-18. It showed good psychometric properties and was sensitive to adolescent concerns. The instrument was designed using ground theory and assesses the following 3 domains associated with quality of life:

Self: Pertains to the adolescents feelings about him or herself.
Relationship: Pertains to the adolescents relations with others.
Environment: Pertains to opportunities and obstacles in the adolescents broader social and cultural milieu.
Co-Authored-By: DonaldPatrick <202009837+DonaldPatrick@users.noreply.github.com>